feat(mcp-client): split core/server modules
