+++ examples/src/main/scala/chimp/adderWithAuthMcp.scala
@@ -4,6 +4,7 @@
 
 package chimp
 
+import chimp.server.*
 import io.circe.Codec
 import sttp.model.Header
 import sttp.tapir.*
