+++ examples/src/main/scala/chimp/twoToolsMcp.scala
@@ -4,6 +4,7 @@
 
 package chimp
 
+import chimp.server.*
 import io.circe.Codec
 import sttp.tapir.*
 import sttp.tapir.server.netty.sync.NettySyncServer
