+++ server/src/main/scala/chimp/server/mcpEndpoint.scala
@@ -1,4 +1,4 @@
-package chimp
+package chimp.server
 
 import io.circe.Json
 import org.slf4j.LoggerFactory
