+++ server/src/test/scala/chimp/server/McpHandlerSpec.scala
@@ -1,4 +1,4 @@
-package chimp
+package chimp.server
 
 import chimp.protocol.*
 import chimp.protocol.JSONRPCMessage.given
